% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSD.R
\name{colSD}
\alias{colSD}
\title{colSD}
\usage{
colSD(data_in)
}
\arguments{
\item{data_in}{matrix of numeric values}
}
\value{
Vector of standard deivations of columns of data_in.
}
\description{
Helper function to calculate standard deviation of matrix by columns
}
\examples{
set.seed(1)
colSD(matrix(rnorm(100), ncol=5))
}
