% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_parameters.R
\name{line_parameters}
\alias{line_parameters}
\title{line_parameters}
\usage{
line_parameters(x1, y1, x2, y2)
}
\arguments{
\item{x1}{first value of the point (x1, x2) in the cartesian plot}

\item{y1}{first value of the point (y1, y2) in the cartesian plot}

\item{x2}{second value of the point (x1, x2) in the cartesian plot}

\item{y2}{second value of the point (y1, y2) in the cartesian plot}
}
\value{
Vector of slope and intercept for equation of line.
}
\description{
Helper function that returns slope and intercept for line
equation using two points in the cartesian plot: (x1, x2) and (y1, y2)
}
\examples{
line_parameters(1,2,4,2)
}
