% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_ae_poisson.R
\name{p_ae_poisson}
\alias{p_ae_poisson}
\title{p_ae_poisson}
\usage{
p_ae_poisson(rate_dc_ae, prob_ae)
}
\arguments{
\item{rate_dc_ae}{Probability of observing at least one AE}

\item{prob_ae}{Proportion of discontinuation due to AE}
}
\value{
Probabilities of discontinuing due to AE.
}
\description{
Helper function that returns probability of discontinuing due
to adverse events (AE)
}
\examples{
p_ae_poisson(c(0.9, 0.8), c(0.1, 0.1))
}
