% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_loe_ee_function.R
\name{p_loe_ee_function}
\alias{p_loe_ee_function}
\title{p_loe_ee_function}
\usage{
p_loe_ee_function(z, p_max, z_l, p_min = 0, z_u, up_good = TRUE)
}
\arguments{
\item{z}{Vector of numeric values, i.e. change from baseline values}

\item{p_max}{Maximum probability of discontinuing}

\item{z_l}{The lower (or left) threshold of the piecewise linear function}

\item{p_min}{Maximum probability of discontinuing (set to 0)}

\item{z_u}{The upper (or right) threshold of the piecewise linear function}

\item{up_good}{TRUE if higher outcome values indicate better responses}
}
\value{
Probabilities of discontinuing due to LoE or EE.
}
\description{
Helper function that returns probability of discontinuing due
to lack of efficacy (LoE) or excess efficacy (EE) via a piecewise linear
function
}
\examples{
line_parameters(1,2,4,2)
}
