% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacf_vec_to_acf.R
\name{pacf_vec_to_acf}
\alias{pacf_vec_to_acf}
\title{pacf_vec_to_acf}
\usage{
pacf_vec_to_acf(pacf_vec, n_repeat)
}
\arguments{
\item{pacf_vec}{Vector of partial autocorrelations}

\item{n_repeat}{number of repeat measures (must be longer than length of pacf_vec)}
}
\value{
Correlation matrix from partial autocorrelations.
}
\description{
Generate correlation matrix from partial autocorrelations
}
\examples{
pacf_vec_to_acf(c(0.5, -0.1), 5)
}
