% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_col.R
\name{rep_col}
\alias{rep_col}
\title{rep_rcol}
\usage{
rep_col(x, n)
}
\arguments{
\item{x}{vector to repeat}

\item{n}{number of repetions}
}
\value{
matrix with vector x repeated n-times by columns.
}
\description{
Helper function to repeat a matrix by column
}
\examples{
set.seed(1)
rep_col(rnorm(5), 5)
}
