% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_row.R
\name{rep_row}
\alias{rep_row}
\title{rep_row}
\usage{
rep_row(x, n)
}
\arguments{
\item{x}{vector to repeat}

\item{n}{number of repetions}
}
\value{
Matrix with vector x repeated n-times by rows.
}
\description{
Helper function to repeat a matrix by row
}
\examples{
set.seed(1)
rep_row(rnorm(5), 5)
}
